+++ components/hubspot/actions/update-deal/update-deal.mjs
@@ -1,13 +1,13 @@
 import { OBJECT_TYPE } from "../../common/constants.mjs";
 import common from "../common/common-update-object.mjs";
-import hubspot from "../../hubspot.app.mjs";
+import appProp from "../common/common-app-prop.mjs";
 
 export default {
   ...common,
   key: "hubspot-update-deal",
   name: "Update Deal",
   description: "Update a deal in Hubspot. [See the documentation](https://developers.hubspot.com/beta-docs/guides/api/crm/objects/deals#update-deals)",
-  version: "0.0.1",
+  version: "0.0.2",
   type: "action",
   methods: {
     ...common.methods,
@@ -16,10 +16,10 @@ export default {
     },
   },
   props: {
-    hubspot,
+    ...appProp.props,
     objectId: {
       propDefinition: [
-        hubspot,
+        appProp.props.hubspot,
         "objectId",
         () => ({
           objectType: "deal",